chore: Update cleanup command to use Redis instead of queue
