+++ app/Jobs/ServerCheckJob.php
@@ -16,7 +16,6 @@ use Illuminate\Contracts\Queue\ShouldBeEncrypted;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
-use Illuminate\Queue\Middleware\WithoutOverlapping;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Arr;
 
@@ -24,7 +23,7 @@ class ServerCheckJob implements ShouldBeEncrypted, ShouldQueue
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
-    public $tries = 3;
+    public $tries = 1;
 
     public $timeout = 60;
 
@@ -44,16 +43,6 @@ class ServerCheckJob implements ShouldBeEncrypted, ShouldQueue
     }
 
     public function __construct(public Server $server) {}
-
-    public function middleware(): array
-    {
-        return [(new WithoutOverlapping($this->server->id))];
-    }
-
-    public function uniqueId(): int
-    {
-        return $this->server->id;
-    }
 
     public function handle()
     {
